+++ SongLib/Core/Di/DiContainer.swift
@@ -25,9 +25,8 @@ final class DiContainer {
             { self.container.resolve(AnalyticsServiceProtocol.self) },
             { self.container.resolve(LoggerProtocol.self) },
             { self.container.resolve(BookDataManager.self) },
-            { self.container.resolve(BookRepositoryProtocol.self) },
             { self.container.resolve(SongDataManager.self) },
-            { self.container.resolve(SongRepositoryProtocol.self) },
+            { self.container.resolve(SongBookRepositoryProtocol.self) },
             { self.container.resolve(SubscriptionRepositoryProtocol.self) },
             { self.container.resolve(ReviewReqRepositoryProtocol.self) },
             { self.container.resolve(SelectionViewModel.self) },
