+++ SongLib/Domain/Repositories/SongBookRepository.swift
@@ -1,5 +1,5 @@
 //
-//  SongRepository.swift
+//  SongBookRepository.swift
 //  SongLib
 //
 //  Created by Siro Daves on 02/05/2025.
@@ -7,29 +7,43 @@
 
 import Foundation
 
-protocol SongRepositoryProtocol {
+protocol SongBookRepositoryProtocol {
+    func fetchRemoteBooks() async throws -> BookResponse
     func fetchRemoteSongs(for bookId: String) async throws -> SongResponse
+    func fetchLocalBooks() -> [Book]
     func fetchLocalSongs() -> [Song]
+    func saveBooks(_ books: [Book])
     func saveSong(_ song: Song)
     func updateSong(_ song: Song)
     func deleteLocalData()
 }
 
-class SongRepository: SongRepositoryProtocol {
-    func updateSong(_ song: Song) {
-        songData.updateSong(song)
-    }
-    
+class SongBookRepository: SongBookRepositoryProtocol {
     private let apiService: ApiServiceProtocol
+    private let bookData: BookDataManager
     private let songData: SongDataManager
     
-    init(apiService: ApiServiceProtocol, songData: SongDataManager) {
+    init(
+        apiService: ApiServiceProtocol,
+        bookData: BookDataManager,
+        songData: SongDataManager
+    ) {
         self.apiService = apiService
+        self.bookData = bookData
         self.songData = songData
+    }
+    
+    func fetchRemoteBooks() async throws -> BookResponse {
+        return try await apiService.fetch(endpoint: .books)
     }
     
     func fetchRemoteSongs(for booksIds: String) async throws -> SongResponse {
         return try await apiService.fetch(endpoint: .songsByBook(booksIds))
+    }
+    
+    func fetchLocalBooks() -> [Book] {
+        let books = bookData.fetchBooks()
+        return books.sorted { $0.bookId < $1.bookId }
     }
     
     func fetchLocalSongs() -> [Song] {
@@ -37,11 +51,20 @@ class SongRepository: SongRepositoryProtocol {
         return songs.sorted { $0.songId < $1.songId }
     }
     
+    func saveBooks(_ books: [Book]) {
+        bookData.saveBooks(books)
+    }
+    
     func saveSong(_ song: Song) {
         songData.saveSong(song)
     }
     
+    func updateSong(_ song: Song) {
+        songData.updateSong(song)
+    }
+    
     func deleteLocalData() {
+        bookData.deleteAllBooks()
         songData.deleteAllSongs()
     }
